Move beam fns to a separate file
